Broke monolithic structure into multiple modular files
