+++ code/sentinalysis/cli.py
@@ -1,0 +1,2 @@
+from utils import check_vader_lexicon
+check_vader_lexicon()
